// src/Utils/types.ts
export type IDType = string | number;

export interface ImageType {
  src?: any;
  alt?: string;
  width?: number;
  height?: number;
  loading?: "lazy" | "eager";
}

export interface HeadingType {
  id?: IDType;
  content?: string;
}

export interface TextType {
  id?: IDType;
  content?: string;
}

export interface ButtonType {
  id?: IDType;
  content?: string;
  path?: string;
  icon?: string;
}